+++ src/pages/MainPage/myComponents/ContentContainer/index.tsx
@@ -1,9 +1,0 @@
-import s from './styles.module.scss'
-
-export default function ContentContainer() {
-    return (
-        <section className={s.content__container}>
-            <p>Unfortunately, I haven't figured out what to put on this page yet, so it's empty for now. 😢</p>
-        </section>
-    )
-}