+++ src/pages/MainPage/myComponents/FindGameButton/index.tsx
@@ -1,9 +1,0 @@
-import s from './styles.module.scss'
-
-export default function FindGameButton() {
-    return (
-        <button className={s.find__game__button}>
-            <p>Find Game</p>
-        </button>
-    )
-}