+++ src/shared/ForFetch/UserAvatar/index.tsx
@@ -11,7 +11,13 @@ type Props = {
 export default function UserAvatar({username, hasFrame, size, frameURL}: Props) {
     return (
         <div className={s.user__avatar__container}>
-            <img className={s.user__avatar} src="/recront.jpg" alt={`${username} Avatar`} draggable={false} style={{width: size + 'px', height: size + 'px'}} />
+            <img
+            className={s.user__avatar}
+            src="/recront.jpg"
+            alt={`${username} Avatar`}
+            draggable={false}
+            style={{width: size, height: size}}
+            />
             {hasFrame && <UserAvatarFrame username={username} frameURL={frameURL} />}
         </div>
     )